WIP: Add a command that uses ImageJ1 data structures
This is a nice demonstration of the fact that it is now possible to
write a parameterized command which works with ImageJ1 images. In this
way, existing ImageJ 1.x plugins can migrate toward the ImageJ2 paradigm
incrementally, making themselves available from other applications
(KNIME, CellProfiler, OMERO, etc.) via the ImageJ2 module
interoperability mechanism -- but without necessarily needing to utilize
the (still in beta) ImageJ2 data model based on ImgLib2.